+++ lib/rubocop/cop/rspec_cops.rb
@@ -22,6 +22,7 @@ rescue LoadError
   # Rails/HttpStatus cannot be loaded if rack/utils is unavailable.
 end
 require_relative 'rspec/rails/inferred_spec_type'
+require_relative 'rspec/rails/response_parsed_body'
 
 require_relative 'rspec/align_left_let_brace'
 require_relative 'rspec/align_right_let_brace'