+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Fix wrong autocorrection in `n_times` style on `RSpec/FactoryBot/CreateList`. ([@r7kamura])
+- Add `RSpec/Rails/ResponseParsedBody` cop. ([@r7kamura])
 
 ## 2.15.0 (2022-11-03)